% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chameleons}
\alias{chameleons}
\title{Occurrence data for Madagascan chameleons}
\format{
An object of class \code{"data.frame"}
}
\source{
(\href{https://onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1111\%2Fj.1558-5646.2008.00596.x&file=EVO_596_sm_AppendixS1.xls}{Appendix S1})
}
\usage{
data(chameleons)
}
\description{
Occurrence data from four species of Madagascan chameleon in the genus \emph{Furcifer}, \emph{F. antimena}, \emph{F. oustaleti}, \emph{F. petteri} and \emph{F. rhinoceratus} taken from Pearson and Raxworthy (2009).
}
\references{
Pearson and Raxworthy (2009) Evolution 63:959-67 doi:10.1111/j.1558-5646.2008.00596.x
(\href{https://pubmed.ncbi.nlm.nih.gov/19210532/}{PubMed})
}
\keyword{data}
