% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{madagascar}
\alias{madagascar}
\title{A RasterLayer defining the extent of Madagascar}
\format{
An object of class \code{"RasterLayer"}
}
\usage{
data(madagascar)
}
\description{
A RasterLayer defining the extent of Madagascar
}
\keyword{data}
